requirements.txt
streamlit>=1.15.0
PyPDF2>=3.0.0
python-docx>=0.8.11